+++ source/src/AgentContourComponent.cs
@@ -12,9 +12,21 @@ namespace RTSCamera
         public uint? Color;
         public bool AlwaysVisible;
     }
+
+    public enum ColorLevel
+    {
+        TargetFormation,
+        SelectedFormation,
+        MouseOverFormation,
+        TargetAgent,
+        SelectedAgent,
+        MouseOverAgent,
+        NumberOfLevel
+    }
+
     public class AgentContourComponent : AgentComponent
     {
-        private readonly Contour[] _colors = new Contour[4];
+        private readonly Contour[] _colors = new Contour[(int)ColorLevel.NumberOfLevel];
         private int _currentLevel = -1;
 
         private uint? CurrentColor => _currentLevel < 0 ? null : _colors[_currentLevel].Color;
@@ -30,19 +42,56 @@ namespace RTSCamera
 
         public void SetContourColor(int level, uint? color, bool alwaysVisible)
         {
-            if (Agent.HasMount)
-                Agent.MountAgent.GetComponent<AgentContourComponent>()?.SetContourColor(level, color, alwaysVisible);
-            if (level < 0 || level >= _colors.Length)
-                return;
-            if (_colors[level].Color == color)
-                return;
-            _colors[level].Color = color;
-            _colors[level].AlwaysVisible = alwaysVisible;
-            if (_currentLevel <= level)
+            if (SetContourColorWithoutUpdate(level, color, alwaysVisible))
             {
                 _currentLevel = color.HasValue ? level : EffectiveLevel(level - 1);
-                Agent.AgentVisuals?.SetContourColor(CurrentColor, CurrentAlwaysVisible);
+                SetColor();
             }
+        }
+
+        public bool SetContourColorWithoutUpdate(int level, uint? color, bool alwaysVisible)
+        {
+            if (level < 0 || level >= _colors.Length)
+                return false;
+            if (_colors[level].Color == color)
+                return false;
+            _colors[level].Color = color;
+            _colors[level].AlwaysVisible = alwaysVisible;
+            return _currentLevel <= level; // needs update.
+        }
+
+        public void UpdateColor()
+        {
+            _currentLevel = EffectiveLevel();
+            SetColor();
+        }
+
+        public void ClearContourColor()
+        {
+            for (int i = 0; i < _colors.Length; ++i)
+            {
+                _colors[i].Color = null;
+                Agent.AgentVisuals?.SetContourColor(null);
+                if (Agent.HasMount)
+                    Agent.MountAgent.AgentVisuals?.SetContourColor(null);
+            }
+        }
+
+        public void ClearTargetOrSelectedFormationColor()
+        {
+            bool needUpdate = SetContourColorWithoutUpdate((int) ColorLevel.TargetFormation, null, true);
+            needUpdate |= SetContourColorWithoutUpdate((int) ColorLevel.SelectedFormation, null, true);
+            if (needUpdate)
+                UpdateColor();
+        }
+
+        public void ClearFormationColor()
+        {
+            bool needUpdate = SetContourColorWithoutUpdate((int)ColorLevel.TargetFormation, null, true);
+            needUpdate |= SetContourColorWithoutUpdate((int)ColorLevel.SelectedFormation, null, true);
+            needUpdate |= SetContourColorWithoutUpdate((int)ColorLevel.MouseOverFormation, null, true);
+            if (needUpdate)
+                UpdateColor();
         }
 
         protected override void OnMount(Agent mount)
@@ -59,7 +108,7 @@ namespace RTSCamera
             mount.AgentVisuals?.SetContourColor(new uint?());
         }
 
-        private int EffectiveLevel(int maxLevel)
+        private int EffectiveLevel(int maxLevel = (int)ColorLevel.NumberOfLevel - 1)
         {
             for (int i = maxLevel; i > -1; --i)
             {
@@ -68,6 +117,13 @@ namespace RTSCamera
             }
 
             return -1;
+        }
+
+        private void SetColor()
+        {
+            Agent.AgentVisuals?.SetContourColor(CurrentColor, CurrentAlwaysVisible);
+            if (Agent.HasMount)
+                Agent.MountAgent.AgentVisuals?.SetContourColor(CurrentColor, CurrentAlwaysVisible);
         }
     }
 }